+++ src/Sky/Canonicalise/Module.hs
@@ -753,7 +753,7 @@ staticKernelFunctions = Map.fromList
                     "andMap", "combine", "traverse", "andThenTask"])
     , ("Task",    ["succeed", "fail", "map", "andThen", "perform", "sequence", "parallel",
                     "lazy", "run", "map2", "map3", "map4", "map5", "andMap",
-                    "fromResult", "andThenResult"])
+                    "fromResult", "andThenResult", "mapError", "onError"])
     , ("Log",     ["println", "debug", "info", "warn", "error", "with", "errorWith"])
     , ("Cmd",     ["none", "batch", "perform"])
     , ("Time",    ["now", "sleep", "every", "unixMillis", "timeString",